rag_chatbot/requirements.txt
openai>=1.0.0
pinecone-client>=3.0.0
streamlit>=1.28.0
yfinance>=0.2.31        
pandas>=2.0.0
numpy>=1.24.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-pinecone>=0.0.1
python-dotenv>=1.0.0
PyPDF2>=3.0.0
requests>=2.31.0        
beautifulsoup4>=4.12.0